Optimizations @ Sun10-05-2020_2-03-45.59

+++ src/com/egangotri/util/EGangotriUtil.groovy
@@ -109,7 +109,7 @@ class EGangotriUtil {
         return hiddenPwd
     }
 
-    static void sleepTimeInSeconds(BigDecimal sleepTimeInSeconds, boolean overrideEgangotriWaitingPeriod = false) {
+    static void sleepTimeInSeconds(double sleepTimeInSeconds, boolean overrideEgangotriWaitingPeriod = false) {
         Thread.sleep(overrideEgangotriWaitingPeriod ? (sleepTimeInSeconds * 1000).toInteger() : (EGangotriUtil.ARCHIVE_WAITING_PERIOD_ONE_SEC * sleepTimeInSeconds).toInteger())
     }
